# docs/swagger.yaml
definitions:
...
  handler.adminTxnListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/model.AdminTransaction'
        type: array
    type: object
...
  handler.blogEnvelope:
    properties:
      data:
        $ref: '#/definitions/model.BlogPost'
    type: object
  handler.blogListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/model.BlogPost'
        type: array
    type: object
  handler.blogRequest:
    properties:
      body:
        type: string
      category:
        type: string
      excerpt:
        type: string
      status:
        example: published
        type: string
      title:
        type: string
    type: object
...
  handler.commentEnvelope:
    properties:
      data:
        $ref: '#/definitions/model.Comment'
    type: object
  handler.commentListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
    type: object
  handler.commentRequest:
    properties:
      body:
        type: string
    type: object
...
  handler.errorEnvelope:
    properties:
      error:
        example: invalid email or password
        type: string
    type: object
...
  handler.loginEnvelope:
    properties:
      data:
        $ref: '#/definitions/handler.loginResponse'
    type: object
  handler.loginRequest:
    properties:
      email:
        example: minhanh@email.com
        type: string
      password:
        example: password
        type: string
    type: object
  handler.loginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
...
  handler.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  handler.revenueEnvelope:
    properties:
      data:
        $ref: '#/definitions/model.RevenueSummary'
    type: object
...
  model.AdminTransaction:
    properties:
      coins:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      kind:
        description: topup | gift
        type: string
      method:
        type: string
      status:
        description: ok | failed
        type: string
      user:
        type: string
      vnd:
        type: integer
    type: object
  model.BlogPost:
    properties:
      author:
        type: string
      body:
        type: string
      category:
        type: string
      comments:
        type: integer
      created_at:
        type: string
      excerpt:
        type: string
      id:
        type: string
      reads:
        type: integer
      status:
        description: published | review | draft
        type: string
      title:
        type: string
    type: object
...
  model.Comment:
    properties:
      author:
        type: string
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      post_id:
        type: string
    type: object
...
  model.RevenueSummary:
    properties:
      month_vnd:
        type: integer
      pro_total:
        type: integer
      today_vnd:
        type: integer
      topups_month:
        type: integer
    type: object
...
  model.User:
    properties:
      coins:
        type: integer
      created_at:
        type: string
      display_name:
        type: string
      elo:
        type: integer
      email:
        type: string
      handle:
        type: string
      id:
        type: string
      plan:
        type: string
      presence:
        type: string
      role:
        description: user | mod | admin
        type: string
      status:
        description: active | banned
        type: string
      status_msg:
        type: string
      streak:
        type: integer
      wins:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/revenue:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.revenueEnvelope'
      security:
      - BearerAuth: []
      summary: Revenue summary (admin)
      tags:
      - admin
...
  /api/v1/admin/transactions:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.adminTxnListEnvelope'
      security:
      - BearerAuth: []
      summary: All transactions (admin, paginated)
      tags:
      - admin
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT valid for 24 hours.
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT and the authenticated user
          schema:
            $ref: '#/definitions/handler.loginEnvelope'
        "400":
          description: malformed request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: invalid email or password
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Log in with email and password
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates an account (if signups are open) and returns a JWT.
      parameters:
      - description: Registration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.loginEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "403":
          description: signups disabled
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "409":
          description: email exists
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Register a new account
      tags:
      - auth
  /api/v1/blog:
    get:
      parameters:
      - description: page (default 1)
        in: query
        name: page
        type: integer
      - description: limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.blogListEnvelope'
      summary: List blog posts (paginated)
      tags:
      - blog
    post:
      consumes:
      - application/json
      parameters:
      - description: Post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/handler.blogRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.blogEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Create a blog post
      tags:
      - blog
  /api/v1/blog/{id}:
    delete:
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: deleted
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete a blog post
      tags:
      - blog
    get:
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.blogEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Get a single blog post
      tags:
      - blog
    put:
      consumes:
      - application/json
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/handler.blogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.blogEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Update a blog post
      tags:
      - blog
  /api/v1/blog/{id}/comments:
    get:
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.commentListEnvelope'
      summary: List comments of a blog post
      tags:
      - blog
    post:
      consumes:
      - application/json
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.commentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.commentEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Add a comment to a blog post
      tags:
      - blog
...
swagger: "2.0"