# internal/handler/blog.go
package handler

import (
	"encoding/json"
	"net/http"

	"github.com/google/uuid"

	"github.com/craftbyte/learning_languages/services/internal/apperror"
	"github.com/craftbyte/learning_languages/services/internal/middleware"
	"github.com/craftbyte/learning_languages/services/internal/model"
	"github.com/craftbyte/learning_languages/services/internal/service"
	"github.com/craftbyte/learning_languages/services/pkg/httputil"
)

type BlogHandler struct {
	blog     service.BlogService
	profiles service.ProfileService
	comments service.CommentService
}

func NewBlogHandler(blog service.BlogService, profiles service.ProfileService, comments service.CommentService) *BlogHandler {
	return &BlogHandler{blog: blog, profiles: profiles, comments: comments}
}

type blogRequest struct {
	Title    string `json:"title"`
	Category string `json:"category"`
	Excerpt  string `json:"excerpt"`
	Body     string `json:"body"`
	Status   string `json:"status" example:"published"`
}

type blogListEnvelope struct { //nolint:unused // referenced by swaggo annotations only
	Data []model.BlogPost `json:"data"`
}

type blogEnvelope struct { //nolint:unused // referenced by swaggo annotations only
	Data model.BlogPost `json:"data"`
}

type commentRequest struct {
	Body string `json:"body"`
}

type commentListEnvelope struct { //nolint:unused // referenced by swaggo annotations only
	Data []model.Comment `json:"data"`
}

type commentEnvelope struct { //nolint:unused // referenced by swaggo annotations only
	Data model.Comment `json:"data"`
}

// ListComments godoc
//
//	@Summary	List comments of a blog post
//	@Tags		blog
//	@Produce	json
//	@Param		id	path		string	true	"Post ID"
//	@Success	200	{object}	commentListEnvelope
//	@Router		/api/v1/blog/{id}/comments [get]
func (h *BlogHandler) ListComments(w http.ResponseWriter, r *http.Request) {
	id, err := uuid.Parse(r.PathValue("id"))
	if err != nil {
		httputil.Error(w, apperror.BadRequest("invalid post id"))
		return
	}
	comments, err := h.comments.List(r.Context(), id)
	if err != nil {
		httputil.Error(w, err)
		return
	}
	httputil.OK(w, comments)
}

// AddComment godoc
//
//	@Summary	Add a comment to a blog post
//	@Tags		blog
//	@Accept		json
//	@Produce	json
//	@Security	BearerAuth
//	@Param		id		path		string			true	"Post ID"
//	@Param		body	body		commentRequest	true	"Comment"
//	@Success	201		{object}	commentEnvelope
//	@Failure	400		{object}	errorEnvelope
//	@Failure	401		{object}	errorEnvelope
//	@Failure	404		{object}	errorEnvelope
//	@Router		/api/v1/blog/{id}/comments [post]
func (h *BlogHandler) AddComment(w http.ResponseWriter, r *http.Request) {
	id, err := uuid.Parse(r.PathValue("id"))
	if err != nil {
		httputil.Error(w, apperror.BadRequest("invalid post id"))
		return
	}
	var req commentRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, apperror.BadRequest("invalid request body"))
		return
	}
	comment, err := h.comments.Create(r.Context(), id, h.authorName(r), req.Body)
	if err != nil {
		httputil.Error(w, err)
		return
	}
	httputil.Created(w, comment)
}

// List godoc
//
//	@Summary	List blog posts (paginated)
//	@Tags		blog
//	@Produce	json
//	@Param		page	query		int	false	"page (default 1)"
//	@Param		limit	query		int	false	"limit (default 20, max 100)"
//	@Success	200	{object}	blogListEnvelope
//	@Router		/api/v1/blog [get]
func (h *BlogHandler) List(w http.ResponseWriter, r *http.Request) {
	page, limit, offset := httputil.PageParams(r)
	posts, total, err := h.blog.List(r.Context(), limit, offset)
	if err != nil {
		httputil.Error(w, err)
		return
	}
	httputil.Paginated(w, posts, page, limit, total)
}

// Get godoc
//
//	@Summary	Get a single blog post
//	@Tags		blog
//	@Produce	json
//	@Param		id	path		string	true	"Post ID"
//	@Success	200	{object}	blogEnvelope
//	@Failure	404	{object}	errorEnvelope
//	@Router		/api/v1/blog/{id} [get]
func (h *BlogHandler) Get(w http.ResponseWriter, r *http.Request) {
	id, err := uuid.Parse(r.PathValue("id"))
	if err != nil {
		httputil.Error(w, apperror.BadRequest("invalid post id"))
		return
	}
	post, err := h.blog.Get(r.Context(), id)
	if err != nil {
		httputil.Error(w, err)
		return
	}
	httputil.OK(w, post)
}

// Create godoc
//
//	@Summary	Create a blog post
//	@Tags		blog
//	@Accept		json
//	@Produce	json
//	@Security	BearerAuth
//	@Param		post	body		blogRequest	true	"Post"
//	@Success	201		{object}	blogEnvelope
//	@Failure	400		{object}	errorEnvelope
//	@Failure	401		{object}	errorEnvelope
//	@Router		/api/v1/blog [post]
func (h *BlogHandler) Create(w http.ResponseWriter, r *http.Request) {
	var req blogRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, apperror.BadRequest("invalid request body"))
		return
	}

	post := &model.BlogPost{
		Title:    req.Title,
		Category: req.Category,
		Author:   h.authorName(r),
		Excerpt:  req.Excerpt,
		Body:     req.Body,
		Status:   req.Status,
	}
	if err := h.blog.Create(r.Context(), post); err != nil {
		httputil.Error(w, err)
		return
	}
	httputil.Created(w, post)
}

// Update godoc
//
//	@Summary	Update a blog post
//	@Tags		blog
//	@Accept		json
//	@Produce	json
//	@Security	BearerAuth
//	@Param		id		path		string		true	"Post ID"
//	@Param		post	body		blogRequest	true	"Post"
//	@Success	200		{object}	blogEnvelope
//	@Failure	400		{object}	errorEnvelope
//	@Failure	401		{object}	errorEnvelope
//	@Failure	404		{object}	errorEnvelope
//	@Router		/api/v1/blog/{id} [put]
func (h *BlogHandler) Update(w http.ResponseWriter, r *http.Request) {
	id, err := uuid.Parse(r.PathValue("id"))
	if err != nil {
		httputil.Error(w, apperror.BadRequest("invalid post id"))
		return
	}
	var req blogRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, apperror.BadRequest("invalid request body"))
		return
	}

	post := &model.BlogPost{
		ID:       id,
		Title:    req.Title,
		Category: req.Category,
		Excerpt:  req.Excerpt,
		Body:     req.Body,
		Status:   req.Status,
	}
	if err := h.blog.Update(r.Context(), post); err != nil {
		httputil.Error(w, err)
		return
	}
	httputil.OK(w, post)
}

// Delete godoc
//
//	@Summary	Delete a blog post
//	@Tags		blog
//	@Produce	json
//	@Security	BearerAuth
//	@Param		id	path	string	true	"Post ID"
//	@Success	204	"deleted"
//	@Failure	401	{object}	errorEnvelope
//	@Failure	404	{object}	errorEnvelope
//	@Router		/api/v1/blog/{id} [delete]
func (h *BlogHandler) Delete(w http.ResponseWriter, r *http.Request) {
	id, err := uuid.Parse(r.PathValue("id"))
	if err != nil {
		httputil.Error(w, apperror.BadRequest("invalid post id"))
		return
	}
	if err := h.blog.Delete(r.Context(), id); err != nil {
		httputil.Error(w, err)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

// authorName resolves the display name of the authenticated author.
func (h *BlogHandler) authorName(r *http.Request) string {
	id, ok := middleware.UserIDFromContext(r.Context())
	if !ok {
		return "Ẩn danh"
	}
	if u, err := h.profiles.Get(r.Context(), id); err == nil {
		return u.DisplayName
	}
	return "Ẩn danh"
}